gitlog-to-changelog: revert inclusion of git-log-fix file
Assaf Gordon reported that this would replace a project's
existing git-log-fix.  This file is always going to be project
specific if it exists at all, so it's better to make it optional.
A corresponding change will be made to GNU hello to improve
usage of the git-log-fix file.

* build-aux/git-log-fix: Delete dummy file.
* modules/gitlog-to-changelog: Don't reference the project
specific git-log-fix file.