gettext: revert "update macros to version 0.19"
This reverts commit 9b9370ca, as it currently requires that
developers of any project that explicitly uses the gettext module
or implicitly uses it through the utimens-tests or
futimens-tests modules, use gettext >= 0.19.
However there are some stability and availablity issues with
that version at present.

We can reinstate this soon, when stability is addressed
and packages are more readily available.