Fix check for pthreads.h pollution on Mingw64
A previous commit attempted to workaround pollution in the
Mingw64 pthreads.h header file

  commit d5fec6c22f03c6a73d62260c9ce091c10c0a9cbd
  Author: Eric Blake <eblake@redhat.com>
  Date:   Wed Jan 22 20:39:45 2014 -0700

    pthread: work around winpthread header pollution on mingw

It activated its workaround based on existance of the strtok_r
wrapper in pthreads.h. As luck would have it, latest Mingw64
headers removed the strtok_r wrapper from pthreads.h, but left
all the others. So the gnulib workaround was not activated.

Change the configure check to look for several of the broken
wrappers, instead of just strtok_r

Signed-off-by: Eric Blake <eblake@redhat.com>