relocatable: Fix that /@unixroot prefix is not working on OS/2 kLIBC
OS/2 kLIBC has a feature to rewrite some path components. For example,
'/@unixroot' is replaced with a value of $UNIXROOT if it is.

So prepending a drive letter to the path starting with '/' makes the
path starting with '/@unixroot' to 'x:/@unixroot' which is unexpected.

This will breaks the behavior of some programs depending on /@unixroot
prefix.

* lib/relocatable.c (relocate): Do not touch pathname if it is started
with '/@unixroot'.