tests: fix 'invalid path dir' error
On OS/2, a path separator is ';' not ':'. And ':' is used as a
separator between a drive letter and directory parts.

As a result, an absolute path such as x:/path/to/dir on OS/2 is
treated as an invalid path dir.

* tests/init.sh (PATH_SEPARATOR): Set at startup.
(path_prepend_): '?:*' is also an absolute path. Use $PATH_SEPARATOR
instead of hard coded ':'.