regex: work around conditional-dependencies glitch
* modules/regex (Depends-on): Add langinfo.
Without this change, I had problems building an experimental
version of GNU Emacs.  The symptom of the bug was a message
‘./configure: line 12726: test: =: unary operator expected’.
This was due to a line in gl_FUNC_NL_LANGINFO that invokes
‘test $HAVE_LANGINFO_CODESET = 1’ even though HAVE_LANGINFO_CODESET
was unset.  Although gl_FUNC_NL_LANGINFO has
‘AC_REQUIRE([gl_LANGINFO_H])’ and gl_LANGINFO_H always sets
HAVE_LANGINFO_CODESET to 0 or 1, gnulib-tool with
--conditional-dependencies sometimes arranges for the
gl_FUNC_NL_LANGINFO code to be executed before the gl_LANGINFO_H
code.  Since the regex code includes <langinfo.h> it should be
depending on the langinfo module anyway, and this happens to work
around the bug, so install that as a workaround for now.  To
reproduce the original problem, run the following shell script on
the version of Gnulib just before this patch was installed.
	rm -fr foo
	mkdir foo
	cat >foo/configure.ac <<'EOF'
	AC_INIT(GNU Emacs, 27.0.50, bug-gnu-emacs@gnu.org, , https://www.gnu.org/software/emacs/)
	gl_EARLY
	gl_INIT
	AC_OUTPUT
	EOF
	./gnulib-tool --import --conditional-dependencies --gnu-make --dir foo regex
	./gnulib-tool --copy build-aux/install-sh foo/install-sh
	./gnulib-tool --copy build-aux/config.sub foo/config.sub
	./gnulib-tool --copy build-aux/config.guess foo/config.guess
	cd foo
	aclocal -I m4
	autoconf
	./configure --with-included-regex